+++ packages/sdk/CHANGELOG.md
@@ -46,6 +46,12 @@
   with `buildAccountSigner`
 - A new optional field `createPlt` to `AuthorizationsV1` which exposes the access structure for PLT creation.
 
+## 10.0.0-alpha.11
+
+### Fixed
+
+- An issue where the token module state of a PLT could not be correctly decoded from it's CBOR representation.
+
 ## 10.0.0-alpha.10
 
 ### Fixed